🚀 [Code] Refactored application structure and updated APIs to .NET 9
- Changed `_favoritesViewModel` and `_shoppingCartViewModel` to `readonly Lazy` in `App.xaml.cs`.
- Updated API base URL in `APIConstants.cs` and modified endpoints.
- Added new namespaces in `GlobalUsings.cs`.
- Changed variable type from `var` to `MauiAppBuilder` in `MauiProgram.cs`.
- Removed `Category.cs` as it is no longer used.
- Updated `Product.cs` to change `Price` type to `float` and modified properties.
- Cleaned up namespaces and redundant code in `MainActivity.cs`, `MainApplication.cs`, `AppDelegate.cs`, and `Program.cs`.
- Updated `tizen-manifest.xml` API version from 8 to 9.
- Changed command name in `launchSettings.json` from "MsixPackage" to "Project".
- Updated `Rantiy.csproj` to target `net9.0` and changed application ID.
- Replaced `appicon.svg`, `appiconfg.svg`, and `splash.svg` with new designs.
- Added `Styles.xaml` for UI component styling.
- Modified `FakeStoreService.cs` to change return type of `GetCategories`.
- Updated `IFakeStoreService.cs` to reflect changes in service methods.
- Changed `HomeViewModel.cs` to use `string` for categories and selected category.
- Updated image handling in `ProductDetailViewModel.cs`.
- Ensured decimal casting for total price in `ShoppingCartViewModel.cs`.
- Updated data binding in `FavoritesPage.xaml` and `FavoritesPage.xaml.cs`.
- Changed image binding in `HomePage.xaml` and `ShoppingCartPage.xaml`.

+++ Rantiy/ViewModels/ShoppingCartViewModel.cs
@@ -8,7 +8,7 @@ public partial class ShoppingCartViewModel : BaseViewModel
 
     public decimal TotalPrice
     {
-        get { return ShoppingCartProducts.Sum(product => product.Price); }
+        get { return (decimal)ShoppingCartProducts.Sum(product => product.Price); }
     }
